pyproject.toml
[project]
name = "pharma-agents"
version = "0.1.0"
description = "Autonomous multi-agent system for molecular property prediction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai[google-genai]>=0.95.0",
    "crewai-tools>=0.17.0",
    "google-genai>=1.0.0",
    "rdkit>=2024.3.1",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.1.1",
    "loguru>=0.7.3",
]

[project.scripts]
pharma-agents = "pharma_agents.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pharma_agents"]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "radon>=6.0.1",
    "ruff>=0.15.6",
    "ty>=0.0.23",
]
